PoC_Version/requirements.txt
arxiv~=2.1.3
beautifulsoup4~=4.12.3
bert_score
chardet~=5.2.0
chromadb~=0.5.3
docx2txt~=0.8
ebooklib~=0.18
elasticsearch
fastapi~=0.115.6
faster_whisper
fire
FlashRank~=0.2.9
fugashi
genanki~=0.13.1
# well fuck gradio. again.
gradio~=5.12.0
html2text~=2024.2.26
jieba~=0.42.1
Jinja2
joblib~=1.4.2
langdetect~=1.0.9
lxml_html_clean
mwparserfromhell~=0.6.6
mwxml~=0.3.4
nltk~=3.9.1
numpy~=1.26.4
onnxruntime~=1.20.1
openai
# Requires rust compiler...
#outlines
pandas~=2.2.2
phonemizer~=3.3.0
Pillow~=10.3.0
#piper-tts
playwright~=1.44.0
# I'm not using postgres yet...
#psycopg2
#psycopg2-binary
psutil~=5.9.8
#psycopg2-binary
pyannote.audio
PyAudio~=0.2.14
pydub~=0.25.1
pymupdf~=1.24.13
pymupdf4llm~=0.0.17
docling~=2.4.0
#PyAudioWPatch # For windows system recording
pypandoc~=1.13
pypandoc_binary
pytest~=8.3.2
python-dotenv~=1.0.1
python-json-logger~=2.0.7
PyYAML~=6.0.1
Requests~=2.32.3
rouge_score~=0.1.2
sacrebleu~=2.4.2
scikit_learn
sentence_transformers
sounddevice # For Handling system audio recording
SQLAlchemy~=2.0.34
streamlit
tenacity~=8.5.0
textstat~=0.7.4
tiktoken~=0.6.0
toml~=0.10.2
tqdm~=4.66.3
trafilatura~=1.9.0
transformers~=4.48.0
urllib3~=2.2.2
yt_dlp
datasets~=2.20.0
tqdm
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple
torch~=2.4.1
torchaudio~=2.4.1
torchvision

bert-score~=0.3.13
loguru~=0.7.3
future~=1.0.0
psycopg2~=2.9.9
psycopg2-binary~=2.9.10
scikit-learn~=1.5.1
sentence-transformers~=3.0.1
bs4~=0.0.2
munch~=4.0.0
scipy~=1.14.1
spaces~=0.30.3
soundfile~=0.12.1
yt-dlp~=2025.1.26
faster-whisper~=1.1.1
keyring~=25.4.1
pycryptodomex~=3.20.0
lxml
jsonlines~=3.1.0
# System audio recording
# pip install sounddevice numpy wave
# brew install blackhole-2ch  # For virtual device
# sudo apt-get install pavucontrol